+++ scm-ui/ui-webapp/src/users/components/navLinks/index.ts
@@ -26,3 +26,4 @@ export { default as EditUserNavLink } from "./EditUserNavLink";
 export { default as SetPasswordNavLink } from "./SetPasswordNavLink";
 export { default as SetPermissionsNavLink } from "./SetPermissionsNavLink";
 export { default as SetPublicKeysNavLink } from "./SetPublicKeysNavLink";
+export { default as SetApiKeysNavLink } from "./SetApiKeysNavLink";