+++ scm-webapp/src/main/java/sonia/scm/api/v2/resources/RepositoryDto.java
@@ -57,6 +57,7 @@ public class RepositoryDto extends HalRepresentation implements CreateRepository
   private String name;
   @NotEmpty
   private String type;
+  private boolean archived;
 
   RepositoryDto(Links links, Embedded embedded) {
     super(links, embedded);