+++ scm-webapp/src/main/java/sonia/scm/repository/HealthChecker.java
@@ -72,10 +72,8 @@ public final class HealthChecker {
   public void checkAll() {
     logger.debug("check health of all repositories");
 
-    PermissionActionCheck<Repository> check = RepositoryPermissions.healthCheck();
-
     for (Repository repository : repositoryManager.getAll()) {
-      if (check.isPermitted(repository)) {
+      if (RepositoryPermissions.healthCheck().isPermitted(repository)) {
         try {
           check(repository);
         } catch (NotFoundException ex) {